title_bar: Factor out application menu into its own component
